+++ src/test/java/bbbnam/체육복/ExersizeWearTest.java
@@ -1,5 +1,6 @@
-package bbbnam;
+package bbbnam.체육복;
 
+import bbbnam.체육복.ExersizeWear;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
